+++ CHANGELOG.md
@@ -30,6 +30,12 @@ Skipping a few digits because what does it really matter?
 
 - Implements a new beta branch, and with it a new `./gravity-sync.sh beta` function to enable it. This will hopefully allow new features and such to be added for test users who can adopt them and provide feedback before rolling out to the main update branch.
 - Uses new SQLITE3 backup methodology introduced in 2.1, for all push/pull sync operations.
+- `./gravity-sync.sh restore` lets you select a different `gravity.db` and `custom.list` for restoration.
+- One new Star Trek reference.
+- `./gravity-sync.sh restore` now shows recent complete Backup executions.
+
+#### 2.1.6
+- Adds prompts during `./gravity-sync.sh configure` to allow custom SSH port and enable PING avoidance.
 
 ## 2.0
 ### The Smart Release